+++ src/com/intellij/idea/plugin/hybris/tools/remote/http/solr/impl/SolrHttpClient.kt
@@ -54,7 +54,7 @@ class SolrHttpClient {
             setAction(CoreAdminParams.CoreAdminAction.STATUS)
             setBasicAuthCredentials(solrConnectionSettings.adminLogin, solrConnectionSettings.adminPassword)
         }
-        .runCatching { process(buildHttpSolrClient(solrConnectionSettings.generatedURL!!)) }
+        .runCatching { process(buildHttpSolrClient(solrConnectionSettings.generatedURL)) }
         .map { parseCoreResponse(it) }
         .getOrElse {
             throw it