+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2024.0.0]
 
+### `Remote Connection` enhancements
+- Redesign HAC connection settings and rely on `PasswordSafe` to store passwords [#968](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/968)
+
 ### Other
 - IntelliJ IDEA API compatibility [#966](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/966)